+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # DSB-FIDE-ID-Registrierung Changelog
 
+## Version 0.1.1 (2023-06-25)
+
+* Delete: Toggle-Icon, da es ersetzt wird durch eine Status-Auswahlliste
+* Add: tl_fideid.status -> Auswahlliste für Bearbeitungsstatus
+* Delete: tl_fideid.antragsteller -> Turnierleiter-Option entfällt, deshalb nicht mehr benötigt
+* Add: tl_fideid.bemerkungen -> Textarea für Bemerkungen vom Formularabsender
+* Add: tl_fideid.antragsteller_ungleich_person -> Optionales Feld, welches bei true die Auftraggeber-Felder einblendet
+
 ## Version 0.1.0 (2023-06-22)
 
 * Add: tl_fideid.formulardatum -> Ungefährer Versandzeitpunkt des Antrags